docs: add install steps for source
This change is necessary to ensure that all users are able
to properly install and use secpat2gf. The new instructions
include the step to run the build command and to re-install
the wheel file to ensure that all dependencies are properly
installed.

By making these changes, users will have a smoother and more
efficient experience when installing and using secpat2gf.
This will ultimately result in better performance and
satisfaction with the tool.