feat: add --only-matching opt to gf flags
The previous grep flags were set to "-aHnPr". However, to
improve the results obtained from the grep function, the
"--only-matching" option has been added to the flags,
resulting in a new set of flags: "-aHnoPr".

This change will result in a more accurate and relevant
output from the grep function.

With this addition, we can now extract only the matching
portion of the line instead of the entire line, making it
easier to analyze and interpret the results.