feat: add support for custom engine
This commit updates the secpat2gf script to provide
more flexibility for users who may want to use a
different engine for processing search patterns.
The script now supports custom engines, in addition
to the default grep engine. See:
https://github.com/tomnomnom/gf#using-custom-engines

The secpat2gf arguments have been updated to
include an option for specifying a custom engine.
The default engine is set to grep, but this can now
be overridden with the `-e` or `--engine` flag.

Additionally, a new field for the engine has been
added to the JSON data generated by the script.
This allows users to easily see which engine was
used for processing their search patterns. The
engine field is set to grep by default, but will
be updated to reflect the custom engine specified
by the user if one is provided.

Overall, this change helps to improve the
flexibility and customization of the secpat2gf
script, and provides users with more control over
their search pattern processing.